Fix various IP address issues
  1.  If the client targets an IPv4 address then redirected request
      could use IPv6.

  2.  Current scanning selects only IPv4 addresses ('A' records), which
      leaves IPv6 addresses untested.

The first item is a limitation of DPM: the IP address is associated with
the redirection; therefore, the client must use the same address family
(IPv4 or IPv6) when redirected.

The second item is a script bug.  Although this is fixed, support for
testing IPv6 addresses is commented out as another (believed to be
unrelated) problem is present.