Update smoke-em.sh and endpoints to support work-arounds
The list of endpoints now includes an additional column, which details
which work-arounds (if any) are needed for this endpoint.  A dash ('-')
indicates that no work-arounds are needed, otherwise each character in
this column indicates a specific work-around.

Support is added in smoke-en.sh script to handle the work-around column
and react to non-dash values.  Initial support is for 'L' to indicate
that the --location-trusted option is needed.

The suffix marker '[*]' is repurposed.  It is now shown when one or more
work-arounds were enabled for this endpoint.