Add SciToken support for token-based direct authn
This patch adds support for using SciTokens when making direct tests of
an endpoint.

Unlike with macaroons, a single SciToken is requested and used for all
operations, and for all IP-addresses.

Success in obtaining a SciToken does NOT increase the endpoints
successful test count.

A failure in obtaining a SciToken does NOT increase the endpoints failed
test count, but it will result in some tests being skipped.

A new debug option is added that includes some additional information.
Currently this is limited to printing the discovered issuing endpoint
and the decoded SciToken.

An X.509 credential is always used when requesting a macaroon from the
third-party server (a macaroon later used to authorise the TPC).  This
is the same behaviour as before.  Some future patch may add support for
requesting the TPC macaroon using a SciToken.

This patch is based on a similar patch by Brian Bockelman