Add work-around for broken TLS 1.3 support in dCache
Motivation:

dCache does not currently work with TLS v1.3 handshakes that involve
(proxy) certificates.  This is described here:

    https://github.com/dCache/dcache/issues/5213

In particular, this requires both dCache running Java 11 and the client
to attempt a TLS v1.3 handshake.

Note.  Although there already exists a work-around in the script for
this issue (see the '-2' command-line option), this is intended only to
apply to interactions with the SUT, not with the third-party ("golden")
endpoint.

The recent upgrade of prometheus.desy.de to Java 11 means that macaroon
requests to authorise third-party operations (such as the data-bearing
GET/PUT for HTTP-TPC) will fail.

Modification:

Adopt the work-around that forces the client to use TLS v1.2.  Unlike
the '-2' work-around, this work-around is always enabled when requesting
a macaroon from the third-party (prometheus, by default).

Result:

Tests should now pass.