Fix bug where direct transfers and file deletion always succeeded.
The checkResult function expects the return-code to be non-zero if there
is a problem.  Currently, if this function is called from the
checkCurlResult function then the return-code is always zero (indicating
no problem) before checkResult is called.

Direct upload, direct download and deletion all use the checkCurlResult
function, so these tests were (mistakenly) always passing.