Fix work-around for the golden endpoint not supporting TLS 1.3
A previous change (in commit 3cfd6e4ac) proported to add a work-around
for the golden endpoint not supporting TLS 1.3.  However, that fix was
incomplete as the third-party push tests also delete the transferred
file (on the golden endpoint), for which the patch failed to introduce
the work-around.

This patch adds the work-around for the two DELETE operations that
target the golden endpoint.