Increase SPEED_TIME, allowing 60 seconds for staled connections
The SPEED_TIME value is the duration over which a specific performance
characteristic must be maintained, otherwise curl will terminate the
operation.  In the smoke tests, this feature is used for direct
transfers: upload or download.

As an example, if a transfer stalls for SPEED_TIME then curl will tear
down the TCP connection and so abort the transfer.

The default value for curl is 30 seconds.

From experience, the current value (15 seconds) is too short -- there
are transfers stalling for 30 seconds that (might) subsequently recover.

A transfer stalling for 15 (or more) seconds is certainly cause for
concern, but it would be better to allow transfers to succeed and
diagnose these issues as performance problems.