concurrency: add initial version of concurrency script
Motivation:

dCache supports HTTP-TPC via the COPY verb.  In order to minimise the
transfer time, dCache should have a steady number of concurrent
transfers.  If the number of concurrent transfers drop then dCache is
starved of activity and cannot keep the pools busy.

Although dCache supports querying the current concurrency (via the admin
interface), what matters is the behaviour over time.  The agent driving
the HTTP-TPC requests must keep dCache busy or risk poor performance.
Instantaneous measurements are both time-consuming (on the dCache admin)
and risk missing the "big picture".

Motification:

Add a python script for parsing the access log file to extract
concurrency information.  This splits a time window into fixed-width
(fixed period) bins.  For each time bin, the code evaluates the minimum,
maximum and average concurrency during that period.  The script outputs
these results for each bin on stdout.  Some progress information is
given on stderr.

The bin width is configurable via a command-line option.  The supplied
value is in seconds, and sub-second resolution is supported.  The
default is five minutes.

Optional exlusion times may be specified.  If specified, these exclude
transfers that lie outside the time range.  The output is also limited
to the specified time window.

Result:

It is now possible to generate data that describes the HTTP-TPC
concurrency dCache is providing.