concurrency: add support for plotting HTTP-TPC concurrency data
Motivation:

The script 'bin/concurrency.py' generates a time-binned tabular data
file that describes the COPY concurrency over time.  Although useful, a
graph would provide a more visual representation, and therefore easier
for humans to understand what is happening.

Modification:

Add a Gnuplot- and bash script to generate a graph that shows HTTP-TPC
concurrency.

The graph shows the minimum and maximum values: both as lines and as a
shaded area.  The average is shown as another line in that graph.

The output uses the SVG format, to allow accurate rendition at whatever
is the display devices resolution.

Result:

It is now possible to generate graphs showing HTTP-TPC concurrency,
using the output from the 'bin/concurrency.py' script.